Moved other Scripts to folder to highlight the new structure
