+++ Not Currently Used/jmath.py
